criação dos endpoints de criação/delete/pegar filmes

+++ mks-case-backend/src/business/movieBusiness.ts
@@ -1,3 +1,4 @@
+import { InvalidToken } from './../error/customError';
 import {
     movieInput, 
     Order,
@@ -9,6 +10,7 @@ import {
 } from "./ports";
 import { MovieDataBase } from "../data/movieDataBase";
 import { MovieNotFound, UnauthorizedUser } from '../error/customError';
+import { FavoriteMovieDataBase } from '../data/favoriteMovieDataBase';
 
 export class MovieBusiness{
     constructor(
@@ -16,10 +18,15 @@ export class MovieBusiness{
         private authenticator: IAuthenticator
     ){ }
 
-    async signup(movie: movieInput):Promise<void>{
+    async input(movie: movieInput):Promise<void>{
         const { title, description, durationInMinutes, token, yearOfRelease } = movie
 
         const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
+
         if(acessToken.role !== 'ADMIN'){
             throw new UnauthorizedUser()
         }
@@ -33,10 +40,15 @@ export class MovieBusiness{
             duration_in_minutes: durationInMinutes,
             year_of_release: yearOfRelease,
         }
-        await new MovieDataBase().signup(input);
+        await new MovieDataBase().input(input);
     }
 
-    async getAll(order: Order):Promise<any>{
+    async getAll(order: Order, token: string):Promise<any>{
+        const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
 
         if(!order || order !== 'ASC' && order !== 'DESC'){
             order = Order.DESC;
@@ -51,6 +63,11 @@ export class MovieBusiness{
         const { title, description, durationInMinutes, token, yearOfRelease, movieId } = movie
 
         const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
+
         if(acessToken.role !== 'ADMIN'){
             throw new UnauthorizedUser()
         }
@@ -67,12 +84,72 @@ export class MovieBusiness{
     }
 
     async delete(id: string, token: string):Promise<void>{
-
         const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
         if(acessToken.role !== 'ADMIN'){
             throw new UnauthorizedUser()
         }
 
+        const checkMovie = await new MovieDataBase().getMovieById(id);
+        if(!checkMovie){
+            throw new MovieNotFound();
+        }
+
         const result = await new MovieDataBase().delete(id);
+    }
+
+    async registerFavMovie(movieId: string, token: string):Promise<void>{
+        const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
+
+        const checkMovie = await new MovieDataBase().getMovieById(movieId);
+        if(!checkMovie){
+            throw new MovieNotFound();
+        }  
+
+        const input = {
+            id: this.idGenerator.generateId(),
+            movie_id: movieId,
+            user_id: acessToken.id
+        }
+
+        const result = await new FavoriteMovieDataBase().register(input);
+    }
+
+    async getFavMovies(token: string):Promise<any>{
+        const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
+
+        const movies = await new FavoriteMovieDataBase().getAllFavoriteMovie(acessToken.id);
+        return movies;
+    }
+
+    async deleteFavMovie(movieId: string, token: string):Promise<void>{
+        const acessToken = this.authenticator.getData(token);
+
+        if(!acessToken.role && !acessToken.id){
+            throw new InvalidToken();
+        }
+
+        const checkMovie = await new MovieDataBase().getMovieById(movieId);
+        if(!checkMovie){
+            throw new MovieNotFound();
+        }  
+
+        const input = {
+            movie_id: movieId,
+            user_id: acessToken.id
+        }
+        
+        const result = await new FavoriteMovieDataBase().delete(input);
     }
 }
